split the unsafe package into multiple modules
